+++ src/main/kotlin/app/revanced/patches/twitch/ad/embedded/annotations/EmbeddedAdsCompatibility.kt
@@ -1,9 +1,10 @@
-package app.revanced.patches.sleepasandroid.annotations
+package app.revanced.patches.twitch.ad.embedded.annotations
 
 import app.revanced.patcher.annotation.Compatibility
 import app.revanced.patcher.annotation.Package
 
-@Compatibility([Package("com.urbandroid.sleep")])
+@Compatibility([Package("tv.twitch.android.app")])
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
-internal annotation class UnlockPremiumCompatibility
+internal annotation class EmbeddedAdsCompatibility
+